clase 7 parte 1: crear carpeta de contextos y abstraer logica
