+++ src/components/BooksList/index.js
@@ -1,11 +1,12 @@
 import { useContext } from "react";
-import { BooksContext, ThemeContext } from "../../App";
+import { BooksContext } from "../../contexts/BooksContext";
+import { ThemeContext } from "../../contexts/ThemeContext";
 import { BookItem, BooksContainer, BookUnorderedList} from "./book-list.styled";
 
 
 const BooksList = () => {
     const books = useContext(BooksContext)
-    const themeStyles = useContext(ThemeContext)
+    const { themeStyles } = useContext(ThemeContext)
 
     return (
         <BooksContainer themeStyles={themeStyles}>
