+++ src/components/Navbar/index.js
@@ -1,10 +1,11 @@
 import { useContext } from "react";
-import { BooksContext, ThemeContext } from "../../App";
+import { BooksContext } from "../../contexts/BooksContext";
+import { ThemeContext } from "../../contexts/ThemeContext";
 import { NavbarContainer } from "./navbar.styled";
 
 export default function Navbar () {
     const books = useContext(BooksContext)
-    const themeStyles = useContext(ThemeContext)
+    const { themeStyles } = useContext(ThemeContext)
     
     return (
            <NavbarContainer themeStyles={themeStyles}>
